requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
rapidfuzz==3.10.0
passlib==1.7.4
argon2-cffi==23.1.0